fix(type-safe-api): fix named composite models and composite model types
Smithy uses the `title` field in OpenAPI to name composite models (all-of, one-of, any-of), which
OpenAPI generator also used to produce more readable generated code. This change updates to use the
title to name composite models if present for parity with OpenAPI generator.

Additionally, nested composite model properties were typed as `any`. Since we know that we always
hoist and generate a type for composite models, we use the model name as the type for these.